When multiple files are processed, adjust orientations of output pages to match input images.
When multiple images are processed and --pagesize is specified, adjust the orientation of each page to match the input image.  Adjustments are based on the orientation of the first page in case the user actually intends to stretch the images.

For example, in a directory with images: 1-portrait.png and 2-landscape.png, img2pdf -s 612x -o output.pdf *.png will result in a pdf where the first page 8.5in x11in and the second page 11in x 8.5in.  Previously, the second page would have been 8.5in x6.568in.