add pyproject.toml
In contrast to setuptools, flit has a hard time with multiple modules.
We could either make the jp2 module a submodule of the img2pdf module or
we could remove the jp2 module entirely. Both options would require
changes for whoever might be using the jp2 module but the former would
also mean a more serious refactoring of the whole codebase. The safer
option is to merge the jp2 module into img2pdf itself.